miner/miner.py
"""
Sample Miner implementation for SN98 ForeverMoney.

This is a reference implementation using a simple rule-based strategy.
Miners can replace this with ML models, optimization algorithms, or hybrid approaches.
"""
import os
import logging
from typing import Dict, Any, List
from flask import Flask, request, jsonify
from dotenv import load_dotenv

from validator.models import (
    ValidatorRequest,
    MinerResponse,
    Strategy,
    Position,
    RebalanceRule,
    MinerMetadata
)
from miner.strategy import SimpleStrategyGenerator

# Load environment
load_dotenv()

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger(__name__)

# Initialize Flask app
app = Flask(__name__)

# Configuration
MINER_VERSION = os.getenv('MINER_VERSION', '1.0.0-mvp')
MODEL_INFO = os.getenv('MODEL_INFO', 'simple-rule-based')
MINER_PORT = int(os.getenv('MINER_PORT', 8000))

# Initialize strategy generator
strategy_generator = SimpleStrategyGenerator()


@app.route('/health', methods=['GET'])
def health_check():
    """Health check endpoint."""
    return jsonify({'status': 'healthy', 'version': MINER_VERSION}), 200


@app.route('/predict_strategy', methods=['POST'])
def predict_strategy():
    """
    Main endpoint for receiving strategy requests from validators.

    Expects JSON payload matching ValidatorRequest schema.
    Returns JSON matching MinerResponse schema.
    """
    try:
        # Parse request
        request_data = request.json
        if not request_data:
            return jsonify({'error': 'Invalid JSON input'}), 400

        # Validate request against schema
        try:
            validator_request = ValidatorRequest(**request_data)
        except Exception as e:
            logger.error(f"Request validation error: {e}")
            return jsonify({'error': f'Invalid request format: {str(e)}'}), 400

        logger.info(
            f"Received strategy request for pair {validator_request.pairAddress}, "
            f"block {validator_request.target_block}"
        )

        # Generate strategy
        strategy = strategy_generator.generate_strategy(validator_request)

        # Construct response
        miner_response = MinerResponse(
            strategy=strategy,
            miner_metadata=MinerMetadata(
                version=MINER_VERSION,
                model_info=MODEL_INFO
            )
        )

        logger.info(
            f"Generated strategy with {len(strategy.positions)} positions"
        )

        return jsonify(miner_response.model_dump()), 200

    except Exception as e:
        logger.error(f"Error processing request: {e}", exc_info=True)
        return jsonify({'error': f'Internal server error: {str(e)}'}), 500


def main():
    """Run the miner server."""
    logger.info(f"Starting SN98 Miner v{MINER_VERSION}")
    logger.info(f"Model: {MODEL_INFO}")
    logger.info(f"Listening on port {MINER_PORT}")

    # Run Flask app
    # In production, use a proper WSGI server like Gunicorn
    app.run(
        host='0.0.0.0',
        port=MINER_PORT,
        debug=False
    )


if __name__ == '__main__':
    main()
